+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - `--no-defaults` flag.
 - XML nodes can be indented with tabs now. Using `--indent=tabs`.
 - Presentational attributes joining can be forced with `--join-style-attributes=all` now.
+- stdin and stdout support.
 
 ### Changed
 - `<FLAG>` argument type accepts `true`, `false`, `yes`, `no`, `y`, `n` now.